requirements.txt
absl-py==2.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.44.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==9.3.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
torch==2.0.1+cpu
torch-tb-profiler==0.4.3
torchvision==0.15.2+cpu
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==3.0.1
